requirements.txt
appdirs==1.4.3
astroid==2.3.1
attrs==19.3.0
autopep8==1.5.4
beautifultable==0.8.0
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.8
csscompressor==0.9.5
defusedxml==0.6.0
Django==2.2.6
django-appconf==1.0.3
django-compressor==2.4
django-extensions==1.9.8
django-filter==2.2.0
django-hashid-field==3.0.0
django-imagekit==4.0.2
django-ipware==2.1.0
django-jalali==3.1.0
django-js-asset==1.2.2
django-modeladmin-reorder==0.3.1
django-oauth-toolkit==1.2.0
django-tinymce==2.8.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
fandogh-cli==1.22.7
gprof2dot==2019.11.30
hashids==1.2.0
idna==2.8
isodate==0.6.0
isort==4.3.21
jdatetime==3.6.1
Jinja2==2.11.2
kavenegar==1.1.2
lazy-object-proxy==1.4.2
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-client==0.0.1
mysqlclient==2.0.1
oauthlib==3.1.0
pilkit==2.0
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.19
Pygments==2.7.1
PyJWT==1.7.1
pylint==2.4.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
rcssmin==1.0.6
requests==2.22.0
requests-toolbelt==0.9.1
retrying==1.3.3
rjsmin==1.1.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
toml==0.10.1
typed-ast==1.4.0
typing==3.7.4.1
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.25.7
websocket-client==0.56.0
wrapt==1.11.2
XlsxWriter==1.2.9
zeep==3.4.0
gunicorn==20.0.4
